Changed a long to an int.  These are the same size in 32-bit Windows and Linux compilers.
git-svn-id: https://svn.code.sf.net/p/jazzplusplus/code/trunk/jazz@960 d001d500-7bcc-4352-ac0f-06d6f3fd1383

+++ src/AsciiMidiFile.cpp
@@ -53,7 +53,7 @@ JZEvent* JZAsciiRead::Read()
 {
   JZEvent* pEvent = 0;
 
-  long Clock;
+  int Clock;
   int StatusByte, Channel, Length;
   mIfs >> Clock >> StatusByte >> Channel >> Length;
   if (mIfs.fail())
